+++ frontend/components/Home/ModuleList.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import Module from "./Module"
+import Module from "./ModuleDisplay/ModuleDisplay"
 import { AllModules_study_modules_with_courses } from "/static/types/moduleTypes"
 
 const moduleColors: Array<{
